requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn[standard]==0.23.1
python-dotenv==1.1.0
openai==1.81.0
python-multipart==0.0.20